# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /orders/{orderId}/submit-for-review:
    put:
      tags:
      - Order
      operationId: submitOrderForReview
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
  /orders/{orderId}/dispute:
    put:
      tags:
      - Order
      operationId: disputeOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
  /orders/{orderId}/cancel:
    put:
      tags:
      - Order
      operationId: cancelOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
  /orders/{orderId}/approve:
    put:
      tags:
      - Order
      operationId: approveOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
  /orders/{orderId}/accept:
    put:
      tags:
      - Order
      operationId: acceptOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: deadlineDate
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
...
  /orders/book/{serviceId}:
    post:
      tags:
      - Order
      operationId: bookService
      parameters:
      - name: serviceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: integer
                format: int64
  /freelancer:
    get:
      tags:
      - Freelancer
      operationId: getFreelancer
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/FreelancerDto'
    post:
      tags:
      - Freelancer
      operationId: createFreelancer
      parameters:
      - name: registrationRequest
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/FreelancerRegistrationRequest'
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/FreelancerDto'
...
  /client:
    get:
      tags:
      - Client
      operationId: getClient
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ClientDto'
    post:
      tags:
      - Client
      operationId: createClient
      parameters:
      - name: registrationRequest
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/ClientRegistrationRequest'
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ClientDto'
...
  /orders:
    get:
      tags:
      - Order
      operationId: getOrders
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 12
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/PageOrderDto'
  /orders/{orderId}:
    get:
      tags:
      - Order
      operationId: getOrder
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/OrderDto'
  /orders/active:
    get:
      tags:
      - Order
      operationId: getActiveOrders
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 12
      responses:
        "404":
          description: Not Found
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "400":
          description: Bad Request
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "500":
          description: Internal Server Error
          content:
            application/hal+json:
              schema:
                oneOf:
                - type: string
                - $ref: '#/components/schemas/ErrorMessage'
        "401":
          description: Unauthorized
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "403":
          description: Forbidden
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/PageOrderDto'
...
components:
  schemas:
    ErrorMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        message:
          type: string
    AddressDto:
      required:
      - city
      - country
      - houseNumber
      - postalCode
      - region
      - street
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        region:
          type: string
        postalCode:
          type: string
        country:
          type: string
        houseNumber:
          type: string
...
    CategoryType:
      type: string
      enum:
      - WEB_DESIGN
      - WEB_DEVELOPMENT
      - FRONTEND_DEVELOPMENT
      - BACKEND_DEVELOPMENT
      - MOBILE_APP_DEVELOPMENT
      - GAME_DEVELOPMENT
      - WORDPRESS_DEVELOPMENT
      - ECOMMERCE_DEVELOPMENT
      - SOFTWARE_DEVELOPMENT
      - API_DEVELOPMENT
      - GRAPHIC_DESIGN
      - UI_UX_DESIGN
      - LOGO_DESIGN
      - BRAND_IDENTITY_DESIGN
      - VIDEO_EDITING
      - ANIMATION
      - ILLISTRATION
      - THREE_D_MODELING
      - PRESENTATION_DESIGN
      - INTERIOR_DESIGN
      - CONTENT_WRITING
      - COPYWRITING
      - GHOSTWRITING
      - TECHNICAL_WRITING
      - RESUME_WRITING
      - GRANT_WRITING
      - CREATIVE_WRITING
      - SEO_WRITING
      - SCRIPTWRITING
      - EDITING_PROOFREADING
      - TRANSLATION
      - TRANSCRIPTION
      - LANGUAGE_TUTORING
      - SUBTITLING_CAPTIONING
      - LOCALIZATION_SERVICES
      - DIGITAL_MARKETING
      - SOCIAL_MEDIA_MARKETING
      - SEO_SEM
      - EMAIL_MARKETING
      - AFFILIATE_MARKETING
      - INFLUENCER_MARKETING
      - LEAD_GENERATION
      - TELEMARKETING
      - BRAND_STRATEGY
      - CUSTOMER_SUPPORT
      - TECHNICAL_SUPPORT
      - VIRTUAL_ASSISTANT
      - CHAT_SUPPORT
      - HELPDESK_SERVICES
      - DATA_ENTRY
      - WEB_RESEARCH
      - PROJECT_MANAGEMENT
      - CALENDAR_MANAGEMENT
      - DOCUMENT_PREPARATION
      - DATA_ANALYSIS
      - DATA_ENGINEERING
      - MACHINE_LEARNING
      - AI_DEVELOPMENT
      - BUSINESS_INTELLIGENCE
      - DATA_VISUALIZATION
      - BIG_DATA_SERVICES
      - CIVIL_ENGINEERING
      - MECHANICAL_ENGINEERING
      - ELECTRICAL_ENGINEERING
      - STRUCTURAL_ENGINEERING
      - CAD_DRAFTING
      - ARCHITECTURE_DESIGN
      - PRODUCT_DESIGN
      - ACCOUNTING
      - BOOKKEEPING
      - FINANCIAL_ANALYSIS
      - BUSINESS_CONSULTING
      - TAX_PREPARATION
      - HR_CONSULTING
      - FINANCIAL_MODELING
      - LEGAL_CONSULTING
      - CONTRACT_DRAFTING
      - INTELLECTUAL_PROPERTY
      - CORPORATE_LAW
      - FAMILY_LAW_SERVICES
      - COACHING_MENTORING
      - FITNESS_TRAINING
      - MUSIC_PRODUCTION
      - PODCAST_EDITING
      - CAREER_COUNSELING
      - EVENT_PLANNING
      - LIFE_COACHING
      - PLUMBING_SERVICES
      - ELECTRICAL_REPAIR
      - CARPENTRY_SERVICES
      - AUTO_REPAIR
      - HOUSE_CLEANING
      - MOVING_HELP
      - GARDENING_LANDSCAPING
      - CONSTRUCTION_WORK
      - PEST_CONTROL
      - LOCKSMITH_SERVICES
      - PAINTING_SERVICES
      - HVAC_REPAIR
      - APPLIANCE_REPAIR
      - FURNITURE_ASSEMBLY
      - PET_SITTING
      - DOG_WALKING
      - BABYSITTING
      - ELDERLY_CARE
      - SECURITY_SERVICES
      - WINDOW_CLEANING
      - WASTE_REMOVAL
      - RECYCLING_SERVICES
      - CAR_WASHING
      - CAR_SCRAPING
      - BIKE_REPAIR
      - FLOORING_INSTALLATION
      - ROOFING_SERVICES
      - WATER_DAMAGE_RESTORATION
      - POOL_CLEANING
      - SNOW_REMOVAL
      - EVENT_STAFFING
      - CATERING_SERVICES
      - PHOTOGRAPHY_SERVICES
      - MAKEUP_ARTIST
      - TATTOO_ARTIST
      - OTHER
...
    LightWeightFreelancerDto:
      required:
      - email
      - firstName
      - id
      - lastName
      - phoneNumber
      - rating
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        rating:
          type: number
          format: double
        email:
          type: string
        phoneNumber:
          type: string
        avatarImageUrl:
          type: string
    LightweightAddressDto:
      required:
      - city
      - country
      - postalCode
      type: object
      properties:
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
    LightweightServiceDto:
      required:
      - categories
      - freelancer
      - id
      - postedAt
      - price
      - rating
      - ratingCount
      - shortDescription
      - thumbnailUrl
      - title
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        shortDescription:
          type: string
        address:
          $ref: '#/components/schemas/LightweightAddressDto'
        rating:
          type: number
          format: double
        ratingCount:
          type: integer
          format: int32
        price:
          type: number
          format: double
        freelancer:
          $ref: '#/components/schemas/LightWeightFreelancerDto'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryType'
        postedAt:
          type: string
          format: date-time
        thumbnailUrl:
          type: string
...
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
    FreelancerRegistrationRequest:
      required:
      - address
      - categories
      - description
      - firstName
      - idDocument
      - lastName
      - phoneNumber
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          $ref: '#/components/schemas/AddressDto'
        phoneNumber:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryType'
        avatarImage:
          type: string
          format: binary
        description:
          type: string
        idDocument:
          type: array
          items:
            type: string
            format: binary
    FreelancerDto:
      required:
      - address
      - email
      - firstName
      - id
      - lastName
      - phoneNumber
      - rating
      - ratingCount
      type: object
      properties:
        id:
          type: integer
          format: int64
        address:
          $ref: '#/components/schemas/AddressDto'
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        rating:
          type: number
          format: double
        ratingCount:
          type: integer
          format: int32
        avatarImageUrl:
          type: string
    ClientRegistrationRequest:
      required:
      - address
      - firstName
      - idDocument
      - lastName
      - phoneNumber
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        address:
          $ref: '#/components/schemas/AddressDto'
        phoneNumber:
          type: string
        avatarImage:
          type: string
          format: binary
        idDocument:
          type: array
          items:
            type: string
            format: binary
    ClientDto:
      required:
      - address
      - email
      - firstName
      - id
      - lastName
      - phoneNumber
      - rating
      - ratingCount
      type: object
      properties:
        id:
          type: integer
          format: int64
        address:
          $ref: '#/components/schemas/AddressDto'
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        rating:
          type: number
          format: double
        ratingCount:
          type: integer
          format: int32
        avatarImageUrl:
          type: string
...
    EventDto:
      required:
      - createdAt
      - id
      - type
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/EventType'
        createdAt:
          type: string
          format: date-time
        disputeId:
          type: integer
          format: int64
...
    LightweightClientDto:
      required:
      - email
      - firstName
      - id
      - lastName
      - phoneNumber
      - rating
      - ratingCount
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        rating:
          type: number
          format: double
        ratingCount:
          type: integer
          format: int32
        email:
          type: string
        phoneNumber:
          type: string
        avatarImageUrl:
          type: string
    LightweightTransactionDto:
      required:
      - id
      - status
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/TransactionStatus'
    OrderDto:
      required:
      - client
      - createdAt
      - events
      - id
      - service
      - status
      - transaction
      type: object
      properties:
        id:
          type: integer
          format: int64
        service:
          $ref: '#/components/schemas/LightweightServiceDto'
        client:
          $ref: '#/components/schemas/LightweightClientDto'
        status:
          $ref: '#/components/schemas/OrderStatus'
        transaction:
          $ref: '#/components/schemas/LightweightTransactionDto'
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventDto'
        additionalNotes:
          type: string
        rejectionReason:
          type: string
        createdAt:
          type: string
          format: date-time
        deadlineDate:
          type: string
          format: date
        updatedAt:
          type: string
          format: date-time
    OrderStatus:
      type: string
      enum:
      - CREATED
      - IN_PROGRESS
      - SUBMITTED_FOR_REVIEW
      - APPROVED
      - COMPLETED
      - DISPUTED
      - CANCELED
    PageOrderDto:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderDto'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
    TransactionStatus:
      type: string
      enum:
      - PENDING
      - ESCROWED
      - RELEASED
      - CANCELED
      - REFUNDED
      - DISPUTED
...